src/pipeline/object_detection.py
import torch
from transformers import ViTImageProcessor, ViTForImageClassification
from PIL import Image

# Load a pre-trained model and processor
# For SIH, you can start with a pre-trained model.
# For production, this would be fine-tuned on custom satellite data.
processor = ViTImageProcessor.from_pretrained("google/vit-base-patch16-224")
model = ViTForImageClassification.from_pretrained("google/vit-base-patch16-224")

def detect_objects(image_path: str) -> dict:
    """
    Detects objects in a satellite image tile using a Vision Transformer.
    
    Args:
        image_path (str): Path to the image tile (e.g., 512x512).

    Returns:
        dict: Model outputs including logits and bounding boxes.
    """
    try:
        image = Image.open(image_path).convert("RGB")
        
        # Preprocess the image
        inputs = processor(images=image, return_tensors="pt")
        
        # Perform inference
        with torch.no_grad():
            outputs = model(**inputs)
            
        # Post-process results
        # For classification, we get logits that can be converted to probabilities
        # In a real satellite image detection system, you'd use a detection model
        # like DETR or a fine-tuned ViT for object detection
        predicted_class_idx = outputs.logits.argmax(-1).item()
        
        # Placeholder for demonstration (simulating detection results)
        results = {
            "predicted_class": predicted_class_idx,
            "confidence": torch.nn.functional.softmax(outputs.logits, dim=-1).max().item(),
            "logits": outputs.logits.tolist(),
            "bboxes": [[100, 100, 150, 150]] # Dummy bbox for demo
        }
        
        print(f"[ObjectDetection] Successfully processed {image_path}")
        return results

    except Exception as e:
        print(f"[Error] Object detection failed: {e}")
        return {}

if __name__ == "__main__":
    # Create a dummy image for testing
    img = Image.new('RGB', (224, 224), color = 'blue')
    img.save("dummy_image.png")
    
    detections = detect_objects("dummy_image.png")
    print("Detections:", detections)
